Using composition instead of specialization for Alarm/Tca info
The use of inheritance/specialization to extend base Notification class
will require the Yang mapping tool to create explicit lists for each of
the concrete sub-classes as well as the super-class. Conditionally
composing the Notification class is a better pattern if it is prefered
to have single list for all types of Notifications.