Issue #353 - Power management for NMC service provisioning
This commit includes extensions to the tapi-photonic-media information model. Specifically, OTSi and Media Channel Service Interface Points (SIPs) are extended with capability information of the operational power range asociated to the related LTPs.
On the other hand, the MC Connectivity Service End-Points (CSEPs) are extended with a set of power management constrains which can be configured at the connectivity service creation to denote intent-like soft constrains.
Moreover, this commit includes an extension for photonic-media links which includes the expected power loss as a state characteristic to be specified by the TAPI server.