OTSI SIP spec #429
This pull request replace current definition of the otsi tunability capabilities as two separate lists: supportable-lower-central-frequency and supportable-upper-central-frequency, by a single supportable-central-frequency-spectrum-band. After discussions into the TAPI calls on18/06/2019 and 25/06/2019, the group agreed on defining the tunabilty capabilities of otsi transmitters as a single list of bands defined by the lower and upper central frequencies of the band, and the frequency-contrains which specify the rest of feasible central frequencies within the band. An OTSi transmitter can have associated more than one band (i.e., C and L bands).