Streaming xmiid clash with Notification impacting UML-Yang tooling
Method of construction of streaming caused three data types in streaming
to have the same xmi id in the .uml file as three data types in
notification. This caused the UML-Yang tooling to build a dependency
between streaming and notification when both were generated together.
This has now been fixed in Streaming. In the process the names of the
types have also been changed to make the visibly distinct in the Yang.
The literal names of the Enumerations have not been changed as these are
only locally relevant in the types.