Yang with if-feature tapi-streaming in streaming augments and connection spec mods
Ensuring that streaming is optional by making the augments depend upon
support for the streaming feature. All yang models in this commit were
enhanced with if-feature.

Also removed the placeholder connection spec class and replaced each
associations that was to the connection spec with simple attribute of
type UUID. These changes were in Connectivity and Topology. The UML was
updated with these changes in the previous push.