Yang version, comment and augmentation changes
Move of all streaming augmentation statements from various models
(equipment, connectivity etc.) to streaming...yang. This removes the
dependency of everything on streaming. Now streaming depends upon everry

Removal of yang-version "1.1" from all files other than streaming. The
version is necessary for the when clause in the augments which is
necessary to prevent structure overloading.

Removal of "compile error" comments in streaming.