Several modifications:
1. TapiEquipment: added PhysicalRouteList (augmenting Connection),
PhysicalRoute and ResiliencePhysicalRoute.

2. TapiNotification: ObjectNotification class, delete the objectContent
attribute, because 1) for bkw compatibility the 2.1.3 Notification model
is still available in 2.3.x, 2) the content of the object is formally
modeled by the augmentations of EventNotification and streaming
LogRecordBody.

3. TapiNotification: ObjectNotification class, change the type of
changedAttributes attribute to simple string (for which the JSON-PATCH
RFC6902 format is prescribed by the RIA 1.1).

4. TapiNotification: Rename ObjectNotification class into
AttributeValueChange class.

5. TapiConnectivity: change ControlHasParameters association from
ExtendedComposite to StrictComposite, similarly to the association
between ConnectivityService and ResilienceConstraint.

6. TapiFm: AlarmCategory enum, removed TCA entry as conceptually wrong.
ITU-T X.733 Event type has different semantic: “the quality of service
alarm type: An alarm of this type is principally associated with a
degradation in the quality of a service.” Not really a TCA.

7. TapiCommon: AlarmConditionName and PmParameterName, added NATIVE
entry.

8. TapiCommon: PmParameterValue is now composed by PmParameterValue,
type Real (which allows either gauges or counters) and PmParameterUnit.

9. TapiFm: AlarmInfo/nativeAlarmInfo from [0..1] to [1].

10. TapiFm: TcaInfo, added nativeTcaInfo and tcaQualifier attributes.

11. TapiFm: AlarmInfo/AlarmQualifier from name/value pairs to simple
string. Similarly for TcaInfo/tcaQualifier,
DetectedCondition/detectedConditionQualifier and
detectedConditionNativeInfo.

12. TapiFm: TcaInfo and PmMetricInfo, added granularityPeriod, optional.

13. TapiFm: TcaInfo/oamJob from [1] to [0..1] for the cases the
measurement job is not visible at the management interface.

14. TapiFm: AlarmInfo, added nativeAlarmName, nativeAlarmInfo from [1]
to [0..1]. TcaInfo, added nativeThresholdIndicatorName, nativeTcaInfo
from [1] to [0..1].

15. TapiOdu: OduPmParameterName, UAS_BID changed to UAS. The “NE/FE/BID”
qualification is already present in OduErrorPerformanceData and is
available in (notified/streamed)
tcaQualifier/detectedConditionQualifier.

16. TapiFm: TcaInfo, added tcaCategory.

17. TapiTopology: added LocalClass to MultiplexingSequence.

18. TapiCommon: introduced the Profile class, composed by TapiContext.

19. TapiPhotonic: only UML, some experimental modifications for optical
impairments.