Breaking out dynamic data from StreamMonitor
The dynamic data has been separated out from the StreamMonitor, into a
separate optional element, such that a shallow get is readily achievable
and a stream/notification of changes can omit the dynamic data such that
the stream is quiet as opposed to extremely noisey.

Comments, conditions and diagram in this area have been improved.