Various enhancements to the equipment model
Correction to typos in various comments. Corrected comments in, or added
comments to, classes, attributes and data types. (e.g., ActualEquipment,
ActualHolder, ActualNonFieldReplaceableModule and
CommonActualProperties).

<<experimental>> stereotype removed from all entities.

General improvements to diagram structure/layout.

Corrections to CONDTIONAL_MANDATORY from MANDATORY along with addition
of a condition for many attributes that are 0..x. Defaults added to many
attributes to assist in construction of condition statements.

Correction of wordings of CONDITIONAL_MANDATORY statements normalized to
“Mandatory where…”.

General correction from RW –-> R (e.g., equipmentNotExpected) and
correction to many stereotypes to WRITE_NOT_ALLOWED (which was broadly
inconsistent with Read Only statement).

isExpectedActualMismatch clarified and made conditional on potential for
expectation.

_accessPort in PhysicalSpan from [2] to [1..*] to allow for multi-ended
spans.

_accessPort removed from PhysicalRoute and replaced with
PhysicalRouteElement.

Correction to splicedStrand (comment added)

Changed comment in Physical Route to “The physical route of a connection
is modeled as an ordered sequence of physical route element instances.
The physical route is a description dedicated to the connection.”

geographicalLocation is now 0..1 as many equipments are fixed by the
containing equipment.

Correction to the definition of equipment to allow expectation and/or
actual statements (but not to represent non-field-replaceable hardware –
the definition was originally derived from the core model and this used
equipment for both roles).

Updated pin and role.

_physicalRoute in PhysicalRouteList is [*] (was [1..*] as a connection
need not have a stated physical route and indeed in most cases does not
have).

Addition of StrandJoint to allow for expression of joint impairments and
reflections via augmentation. FlowDirection has been enhancement from
earlier discussions to allow for efficient expression of bidirectional
impairments and bidirectional reflections to avoid having to restate
values.

localId added to physical route.

Physical route made optional (as not all connections have physical
route).

connectorPinOrientation added to PinAndRole.

Addition of equipmentNotExpected property in ExpectedEquipment to allow
differentiation between the case of an explicit expectation of nothing
present from no expectation.

Change to multiplicity of commonEquipmentProperties in ExpectedEquipment
to 0..1.

PhyisicalRouteElementHasAccessPort association multiplicity corrected to
0..1 at both ends to allow for case where the route is expressed via
connector-pin alone.

Comment on ConnectorPinOnEquipment end _equipment corrected to include
both connector and pin.

Comments removed from the diagram (the essence of each comment has been
moved to the appropriate class).

AbstractStrand in PhysicalSpan changed from [1..*] to [*] to allow for
extremely lightweight model of the PhysicalSpan.

AccessPort augmentation of SIP changed to be independent of equivalent
for NEP as comments and labels were all for NEP and adjusting these
proved challenging. Name changed/introduced AccessPortSupportingNep and
AccessPortSupportingSip.

ConnectorPinInOnEquipment… “InOn” typo corrected.