Adding qualifiedMeasurementType
This allows for multi-type measurements. In conjunction with
qualifiedValueName and qualifiedLocationName the solution will deal with
cases where there is more than one detector of the same type (i.e., on
more than one signal in the TP), where that detector measures more than
one aspect of the signal and an aspect can have min, max and mean
values.