<<ExtendedComposite>>, NOT_APPLICABLE clash and isLeaf true
Augmentations replaced by <<ExtendedComposite>> as augmentations were
introducing unnecessary layers.

NOT_APPLICABLE in DirectionOfMeasuredSignal replaced with NO_DIRECTION
(some protobuf tooling was complaining about use of NOT_APPLICABLE label
in 2 different Enumerations - not clear why!).

isLeaf true in all enumerations as these are NOT identities.
Enumerations in this model are unlikely to need extension. Assumption is
that extension will be carried out if necessary, from release to
release, by adding to the end of the list and ensuring that no values
are ever removed (simply deprecated). Extension by external parties is
not recommended (as there will be enumeration clashes).