Final adjustments to topology min-elements.
The specific issue is that the UML to YANG tool codes a specific structure in UML to uses in YANG. The structure in UML has a natural optionality such that a mandatory element below this structure is not mandatory overall. This indirection is not present in the YANG.

The cases of this "uses" problem in the topology model have all had the min-elements removed (from tapi-topology.yang):

- layer-protocol-transition-pac / transitioned-layer-protocol-name (min-elements 2;)

- risk-parameter-pac / risk-characteristic (min-elements 1)

- risk-characteristic / risk-identifier-list (min-elements 1)

- transfer-cost-pac /cost-characteristic (min-elements 1)

- transfer-timing-pac / latency-characteristic (min-elements 1)

- validation-pac / validation-mechanism (min-elements 1)

The UML has also been updated (converting 2..* and 1..* to 0..*).