Enabling link properties to be modified to enable TR-547 UC 18a/b
Properties in the Link need to be writable. This is covered by issue #604. The use case identifies
- name
- administrative-state
- cost-characteristic
- latency-characteristic
- risk-characteristic
- transfer-integrity

These were effectively blocked by config-false statements at topology and link. These constraints have been removed.

There may be some side effects that cause other things to seem to be writable. This has been minimised. This commit covers the YANG. No changes are required to the OAS or RIA. The UML is still to be assessed.